Change Classes processing
