+++ default/python/stub_generator/generate_stub.py
@@ -1,39 +1,32 @@
 from logging import error, warning
 from operator import attrgetter, itemgetter
-from typing import List
+from typing import Dict, Iterable, List, Set
 
 from common.print_utils import Table, Text
 from stub_generator.constants import ATTRS, CLASS_NAME, DOC, ENUM_PAIRS, NAME, PARENTS
-from stub_generator.interface_inspector import EnumInfo, FunctionInfo
+from stub_generator.interface_inspector import ClassInfo, EnumInfo, FunctionInfo
 from stub_generator.parse_docs import Docs
 
 
-def _handle_class(info):
-    name = info[NAME]
-    docs = info[DOC]
-    attrs = info[ATTRS]
-    assert not docs, "Got docs need to handle it"
-    parents = info[PARENTS]
-    if not parents:
-        parents = ['object']  # instance is boost wrapper
-    result = []
+def _handle_class(info: ClassInfo):
+    assert not info.doc, "Got docs need to handle it"
+    parents = [x for x in info.parents if x != 'object']
 
-    if 'object' in parents:
-        parents.remove('object')
+    result = []
     if parents:
-        result.append('class %s(%s):' % (name, ', '.join(parents)))
+        result.append('class %s(%s):' % (info.name, ', '.join(info.parents)))
     else:
-        result.append('class %s:' % name)
+        result.append('class %s:' % info.name)
 
     properties = []
     instance_methods = []
-    for attr_name, attr in sorted(attrs.items()):
+    for attr_name, attr in sorted(info.attributes.items()):
         if attr['type'] == "<class 'property'>":
             properties.append((attr_name, attr.get('rtype', '')))
         elif attr['type'] in ("<class 'Boost.Python.function'>", "<class 'function'>"):
             instance_methods.append(attr['routine'])
         else:
-            warning("Skipping '%s' (%s): %s" % (name, attr['type'], attr))
+            warning("Skipping '%s' (%s): %s" % ((info.name), attr['type'], attr))
 
     for property_name, rtype in properties:
         if not rtype:
@@ -116,7 +109,7 @@ def _handle_enum(info: EnumInfo):
     return '\n'.join(result)
 
 
-def _report_classes_without_instances(classes_map, instance_names, classes_to_ignore: set):
+def _report_classes_without_instances(classes_map: Iterable[str], instance_names, classes_to_ignore: Set[str]):
     missed_instances = instance_names.symmetric_difference(classes_map).difference(classes_to_ignore)
 
     if not missed_instances:
@@ -137,14 +130,14 @@ def _report_classes_without_instances(classes_map, instance_names, classes_to_ig
     warning(table.get_table())
 
 
-def make_stub(classes, enums: List[EnumInfo], functions: List[FunctionInfo], instances, result_path, classes_to_ignore: set):
+def make_stub(classes: List[ClassInfo], enums: List[EnumInfo], functions: List[FunctionInfo], instances, result_path, classes_to_ignore: set):
 
     # exclude technical Map classes that are prefixed with map_indexing_suite_ classes
-    classes = [x for x in classes if not x[NAME].startswith('map_indexing_suite_')]
-    classes_map = {x[NAME]: x for x in classes}
+    classes = [x for x in classes if not x.name.startswith('map_indexing_suite_')]
+    classes_map = {x.name: x for x in classes}
 
     _report_classes_without_instances(
-        classes_map,
+        classes_map.keys(),
         {instance[CLASS_NAME] for instance in instances},
         classes_to_ignore)
 
@@ -159,7 +152,7 @@ def make_stub(classes, enums: List[EnumInfo], functions: List[FunctionInfo], ins
             warning("skipping enum instance: %s" % class_name)
             continue
         try:
-            class_attrs = classes_map[class_name][ATTRS]
+            class_attrs = classes_map[class_name].attributes
         except KeyError:
             error("Instance class was not found in classes generated by C++ interface,"
                   " please check instance at %s with class: %s" % (instance["location"], class_name,))
@@ -192,8 +185,8 @@ def make_stub(classes, enums: List[EnumInfo], functions: List[FunctionInfo], ins
         'from typing import Dict'
     ]
     classes = sorted(classes, key=lambda class_: (
-        len(class_[PARENTS]), class_[PARENTS] and class_[PARENTS][0] or '',
-        class_[NAME]))  # put classes with no parents on first place
+        len(class_.parents), class_.parents and class_.parents[0] or '',
+        class_.name))  # put classes with no parents on first place
 
     for cls in classes:
         res.append(_handle_class(cls))
