+++ default/python/stub_generator/interface_inspector/__init__.py
@@ -1,3 +1,4 @@
 from stub_generator.interface_inspector.module_inspector import get_module_info
 from stub_generator.interface_inspector.enum_processor import EnumInfo
 from stub_generator.interface_inspector.function_processor import FunctionInfo
+from stub_generator.interface_inspector.class_processor import ClassInfo
